+++ plugin/src/java/nl/esciencecenter/e3dchem/gpcrdb/GpcrdbNodeModel.java
@@ -1,33 +1,42 @@
 package nl.esciencecenter.e3dchem.gpcrdb;
 
+import java.util.concurrent.TimeUnit;
+
 import org.knime.core.node.InvalidSettingsException;
 import org.knime.core.node.NodeModel;
 import org.knime.core.node.NodeSettingsRO;
 import org.knime.core.node.NodeSettingsWO;
+import org.knime.core.node.defaultnodesettings.SettingsModelInteger;
 import org.knime.core.node.defaultnodesettings.SettingsModelString;
 
 import nl.esciencecenter.e3dchem.gpcrdb.client.ApiClient;
 
-
 public abstract class GpcrdbNodeModel extends NodeModel {
 
 	public static final String CFGKEY_BASEPATH = "Base path";
-
 	public static final String DEFAULT_BASEPATH = "http://gpcrdb.org/";
-
 	private final SettingsModelString m_basePath = basePathSettings();
+
+	public static final String CFGKEY_TIMEOUT = "Timeout";
+	public static final int DEFAULT_TIMEOUT = 10;
+	private final SettingsModelInteger m_timeout = timeoutSettings();
 
 	private ApiClient apiClient;
 
 	public static SettingsModelString basePathSettings() {
-		return new SettingsModelString(CFGKEY_BASEPATH,	DEFAULT_BASEPATH);
+		return new SettingsModelString(CFGKEY_BASEPATH, DEFAULT_BASEPATH);
 	}
-	
+
+	public static SettingsModelInteger timeoutSettings() {
+		return new SettingsModelInteger(CFGKEY_TIMEOUT, DEFAULT_TIMEOUT);
+	}
+
 	protected GpcrdbNodeModel(int nrInDataPorts, int nrOutDataPorts) {
 		super(nrInDataPorts, nrOutDataPorts);
 		apiClient = new ApiClient();
 		apiClient.setBasePath(m_basePath.getStringValue());
 		apiClient.addDefaultHeader("Accept", "application/json");
+		apiClient.getHttpClient().setReadTimeout(m_timeout.getIntValue(), TimeUnit.SECONDS);
 	}
 
 	/**
@@ -45,6 +54,7 @@ public abstract class GpcrdbNodeModel extends NodeModel {
 	protected void loadValidatedSettingsFrom(final NodeSettingsRO settings) throws InvalidSettingsException {
 		m_basePath.loadSettingsFrom(settings);
 		apiClient.setBasePath(m_basePath.getStringValue());
+		apiClient.getHttpClient().setReadTimeout(m_timeout.getIntValue(), TimeUnit.SECONDS);
 	}
 
 	/**
@@ -62,14 +72,20 @@ public abstract class GpcrdbNodeModel extends NodeModel {
 	public void setApiClient(ApiClient apiClient) {
 		this.apiClient = apiClient;
 	}
-	
+
 	public String getBasePath() {
 		return m_basePath.getStringValue();
 	}
-	
+
 	public void setBasePath(String basePath) {
 		m_basePath.setStringValue(basePath);
 	}
+
+	public int getTimeout() {
+		return m_timeout.getIntValue();
+	}
+
+	public void setTimeout(int timeout) {
+		m_timeout.setIntValue(timeout);
+	}
 }
-
-