+++ plugin/src/java/nl/esciencecenter/e3dchem/gpcrdb/mutants/MutantsNodeDialog.java
@@ -32,6 +32,6 @@ public class MutantsNodeDialog extends GpcrdbNodeDialog {
 
 		createNewTab("Advanced");
 
-		addBasePath();
+		addAdvancedSettings();
 	}
 }