+++ plugin/src/java/nl/esciencecenter/e3dchem/gpcrdb/proteins/ProteinInfoNodeDialog.java
@@ -36,6 +36,6 @@ public class ProteinInfoNodeDialog extends GpcrdbNodeDialog {
 
 		createNewTab("Advanced");
 
-		addBasePath();
+		addAdvancedSettings();
 	}
 }