+++ plugin/src/java/nl/esciencecenter/e3dchem/gpcrdb/proteins/ProteinSimilarityNodeDialog.java
@@ -49,6 +49,6 @@ public class ProteinSimilarityNodeDialog extends GpcrdbNodeDialog {
         addDialogComponent(new DialogComponentNumber(
                 new SettingsModelIntegerBounded(ProteinSimilarityNodeModel.CFGKEY_CHUNKSIZE, 20, 1, 100), "Chunk Size", 1));
 
-        addBasePath();
+        addAdvancedSettings();
     }
 }