+++ plugin/src/java/nl/esciencecenter/e3dchem/gpcrdb/residues/ResiduesNodeDialog.java
@@ -36,6 +36,6 @@ public class ResiduesNodeDialog extends GpcrdbNodeDialog {
 
 		createNewTab("Advanced");
 
-		addBasePath();
+		addAdvancedSettings();
 	}
 }