+++ plugin/src/java/nl/esciencecenter/e3dchem/gpcrdb/structures/InteractionsNodeDialog.java
@@ -31,6 +31,6 @@ public class InteractionsNodeDialog extends GpcrdbNodeDialog {
 
 		createNewTab("Advanced");
 
-		addBasePath();
+		addAdvancedSettings();
 	}
 }