+++ tests/src/java/nl/esciencecenter/e3dchem/gpcrdb/residues/ResiduesNodeModelTest.java
@@ -1,5 +1,6 @@
 package nl.esciencecenter.e3dchem.gpcrdb.residues;
 
+import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -85,4 +86,9 @@ public class ResiduesNodeModelTest {
 		return response;
 	}
 
+	@Test
+	public void test_defaultTimeout() {
+		ResiduesNodeModel model = new ResiduesNodeModel();
+		assertEquals(model.getTimeout(), 10);
+	}
 }